+++ app/src/main/java/com/example/helplineapp/MainActivity.kt
@@ -7,16 +7,11 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.annotation.RequiresApi
-import androidx.core.view.WindowInsetsControllerCompat
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.example.helplineapp.View.Forum.ForumScreen
-import com.example.helplineapp.View.Login.LoginScreen
-import com.example.helplineapp.View.NavDrawer
 import com.example.helplineapp.ui.theme.HelplineAppTheme
-import com.example.helplineapp.View.Notification.NotificationScreen
-import com.example.helplineapp.View.SplashScreen.SplashScreen
+import com.example.helplineapp.view.vaga.VagaScreen
 
 class MainActivity : ComponentActivity() {
   @SuppressLint("WrongConstant")
@@ -24,35 +19,40 @@ class MainActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
     enableEdgeToEdge()
-    WindowInsetsControllerCompat(window, window.decorView).let { controller ->
-      controller.hide(android.view.WindowInsets.Type.systemBars())
-      controller.systemBarsBehavior =
-        WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
-    }
+//    WindowInsetsControllerCompat(window, window.decorView).let { controller ->
+//      controller.hide(android.view.WindowInsets.Type.systemBars())
+//      controller.systemBarsBehavior =
+//        WindowInsetsControllerCompat.BEHAVIOR_SHOW_TRANSIENT_BARS_BY_SWIPE
+//    }
     setContent {
       HelplineAppTheme {
         val navController = rememberNavController()
 
         // startDestination -> Tela que o aplicativo vai começar
-        NavHost(navController = navController, startDestination = "splashScreen") {
+        NavHost(navController = navController, startDestination = "VagaScreen") {
 
-          // Criando a rota para a tela de login
-          composable(route = "loginPage") {
-            LoginScreen(navController)
-          }
+//          // Criando a rota para a tela de login
+//          composable(route = "loginPage") {
+//            LoginScreen(navController)
+//          }
+//
+//          // Rota para tela do fórum
+//          composable(route = "forumScreen") {
+//            ForumScreen(navController)
+//          }
+//
+//          // Rota para tela de splash screen
+//          composable(route = "splashScreen") {
+//            SplashScreen(navController)
+//          }
+//
+//          composable(route = "notificationScreen") {
+//            NotificationScreen(navController)
+//          }
 
-          // Rota para tela do fórum
-          composable(route = "forumScreen") {
-            ForumScreen(navController)
-          }
-
-          // Rota para tela de splash screen
-          composable(route = "splashScreen") {
-            SplashScreen(navController)
-          }
-
-          composable(route = "notificationScreen") {
-            NotificationScreen(navController)
+          // Rota para tela da vaga
+          composable(route = "VagaScreen"){
+            VagaScreen(navController)
           }
         }
       }
