+++ app/src/main/java/com/example/helplineapp/view/Cadastro/CadastroScreen.kt
@@ -1,0 +1,147 @@
+package com.example.helplineapp.view.Cadastro
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.sp
+import com.example.helplineapp.R // Certifique-se de usar o R do seu pacote
+
+@Composable
+fun CadastroScreen() {
+    var email by remember { mutableStateOf("") }
+    var confirmEmail by remember { mutableStateOf("") }
+    var senha by remember { mutableStateOf("") }
+    var confirmSenha by remember { mutableStateOf("") }
+
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(Color(0xFFFAF9F7))
+    ) {
+        Image(
+            painter = painterResource(id = R.drawable.background_image),
+            contentDescription = null,
+            modifier = Modifier.fillMaxSize(),
+            contentScale = ContentScale.Crop
+        )
+
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = 48.dp, vertical = 44.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Image(
+                painter = painterResource(id = R.drawable.logo),
+                contentDescription = "Logo",
+                modifier = Modifier
+                    .width(163.dp)
+                    .aspectRatio(1.43f)
+            )
+
+            Spacer(modifier = Modifier.height(44.dp))
+
+            Text(
+                text = "Informações de login",
+                modifier = Modifier.fillMaxWidth(),
+                textAlign = TextAlign.Center,
+                style = TextStyle(
+                    fontSize = 24.sp,
+                    fontWeight = FontWeight.Bold
+                )
+            )
+
+            Column(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(1.dp),
+                verticalArrangement = Arrangement.Center,
+                horizontalAlignment = Alignment.CenterHorizontally
+            ) {
+                TextField(
+                    value = email,
+                    onValueChange = { email = it },
+                    label = { Text("E-mail") },
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(4.dp),
+                    shape = RoundedCornerShape(20.dp),
+                    singleLine = true
+                )
+                Spacer(modifier = Modifier.height(16.dp))
+
+                TextField(
+                    value = confirmEmail,
+                    onValueChange = { confirmEmail = it },
+                    label = { Text("Confirmar E-mail") },
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(4.dp),
+                    shape = RoundedCornerShape(20.dp),
+                    singleLine = true
+                )
+                Spacer(modifier = Modifier.height(16.dp))
+
+                TextField(
+                    value = senha,
+                    onValueChange = { senha = it },
+                    label = { Text("Senha") },
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(4.dp),
+                    shape = RoundedCornerShape(20.dp),
+                    singleLine = true,
+                    visualTransformation = PasswordVisualTransformation()
+                )
+                Spacer(modifier = Modifier.height(16.dp))
+
+                TextField(
+                    value = confirmSenha,
+                    onValueChange = { confirmSenha = it },
+                    label = { Text("Confirmar Senha") },
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(4.dp),
+                    shape = RoundedCornerShape(20.dp),
+                    singleLine = true,
+                    visualTransformation = PasswordVisualTransformation()
+                )
+                Spacer(modifier = Modifier.height(32.dp))
+
+                Button(
+                    onClick = { /* lógica de validação e navegação aqui */ },
+                    modifier = Modifier
+                        .width(200.dp)
+                        .height(50.dp)
+                        .wrapContentSize(align = Alignment.Center),
+                    colors = ButtonDefaults.buttonColors(containerColor = Color(28, 54, 30)),
+                    shape = RoundedCornerShape(15.dp)
+                ) {
+                    Text("Próximo", color = Color.White)
+                }
+            }
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun CadastroScreenPreview() {
+    CadastroScreen()
+}
+
+//Teste
